lab6/lab6_2/src/main/java/tqs/lab3_2/Service/CarManagerService.java
package tqs.lab3_2.Service;

import java.util.List;
import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import tqs.lab3_2.data.CarRepository;


@Service
public class CarManagerService {

    @Autowired
    private CarRepository carRepository;

    public Car save(Car car){
        return this.carRepository.save(car);
    }

    public List<Car> getAllCars(){
        return this.carRepository.findAll();
    }

    public Optional<Car> getCarDetails(Long carld){
        return this.carRepository.findByCarld(carld);
    }

}
